Move templates into their own directories
