+++ client/src/components/event-template/bracket-segment.js
